Finished adding new system for calculting balance.
System should work perfectly, no bugs in testing so I think it is ready for pull request.

With the new system, I should be able to clean up some of the other code but I can't be bothered doing that now.

Also may need some more comments, but I think I got most of it.